+++ src/context/Search/useSearch.test.ts
@@ -42,7 +42,7 @@ const goodFetchMockImplementation = () => {
 };
 
 const reportError = jest.fn();
-jest.mock("../../common-lib/error-reporter", () => ({
+jest.mock("@/common-lib/error-reporter/errorReporter", () => ({
   __esModule: true,
   default:
     () =>
@@ -161,7 +161,7 @@ describe("useSearch()", () => {
 
     await waitFor(() => expect(result.current.status).toBe("fail"));
   });
-  test.skip("error should be reported", async () => {
+  test("error should be reported", async () => {
     // @todo skipping this test, not sure why it's failing
     const error = new Error("bad thing");
     fetch.mockRejectedValue(error);